% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symbols_entrezids2biopax.R
\name{add_symbols_entrezids2biopax}
\alias{add_symbols_entrezids2biopax}
\title{Annotate a BioPAX Object with Gene ID and Gene Symbol}
\usage{
add_symbols_entrezids2biopax(biopax)
}
\arguments{
\item{biopax}{A BioPAX object.}
}
\description{
Adds Gene IDs and Symbols to all physical entities, where possible.
}
\author{
Ivan Grishagin
}
