% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symbols_entrezids_mult_keytypes.R
\name{add_symbols_entrezids_mult_keytypes}
\alias{add_symbols_entrezids_mult_keytypes}
\title{Annotate DataFrame with Gene ID and Gene Symbol}
\usage{
add_symbols_entrezids_mult_keytypes(dFrame, col_names = c(id_col =
  "biopax.Gene.ID", type_col = "biopax.Gene.ID.Type", entrez_col =
  "ENTREZID", symbol_col = "biopax.Gene.Symbol", pathway_col =
  "biopax.Pathway.ID", comp_col = "biopax.Component.ID"),
  keytypes = c("entrezgene", "mim", "uniprot", "unigene", "ensemblgene",
  "hprd", "hgnc"), filter_keytypes = TRUE)
}
\arguments{
\item{dFrame}{Dataframe or data table to process.}

\item{col_names}{A named vector of actual column names in the supplied \code{dFrame}.}

\item{keytypes}{A format of IDs to be converted.}

\item{filter_keytypes}{Logical. Return only rows with desired format types? Defaults to \code{TRUE}.}
}
\description{
Adds a Gene ID and Symbol columns to a dataframe with a protein/gene ids in multiple other formats.
}
\author{
Ivan Grishagin
}
