% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symbols_entrezids_single_keytype.R
\name{add_symbols_entrezids_single_keytype}
\alias{add_symbols_entrezids_single_keytype}
\title{Annotate DataFrame with Gene ID and Gene Symbol}
\usage{
add_symbols_entrezids_single_keytype(dFrame, col_names = c(id_col =
  "id_col", type_col = "type_col", entrez_col = "entrez_col", symbol_col =
  "symbol_col", pathway_col = "", comp_col = ""), KEYTYPE = NULL,
  species = "human")
}
\arguments{
\item{dFrame}{Dataframe or data table to process.}

\item{col_names}{A named vector of actual column names 
in the supplied \code{dFrame}.}

\item{KEYTYPE}{A format of IDs to be converted.}

\item{species}{Species to look for. Default is "human".}
}
\description{
Adds a Gene ID and Symbol columns to a dataframe with a protein/gene ids 
in one other format (see \code{KEYTYPE} argument).
}
\author{
Ivan Grishagin
}
