% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pathways_from_biopax.R
\name{extract_pathways_from_biopax}
\alias{extract_pathways_from_biopax}
\title{Extract Pathways from BioPAX}
\usage{
extract_pathways_from_biopax(biopax, pw_ids, filename = NULL,
  exclude_subpw_comp_patt = NULL)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pw_ids}{Pathway ids to extract.}

\item{filename}{Filename of the new BioPAX}

\item{exclude_subpw_comp_patt}{Pattern of sub-pathways, whose components will be excluded, 
i.e. the main pathway will refer to them, but the references to their components will be removed.}
}
\description{
Extract pathways from a BioPAX object, then convert them into a separate biopax object.
}
\author{
Ivan Grishagin
}
