% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netpath_owls.R
\name{get_netpath_owls}
\alias{get_netpath_owls}
\title{Download NetPath BioPAX Files}
\usage{
get_netpath_owls(dir = getwd())
}
\arguments{
\item{dir}{Directory, where the files will be saved to in a new subdirectory \code{NetPath_owls}.}
}
\description{
Downloads NetPath BioPAX files in the *.owl format.
}
\author{
Ivan Grishagin
}
