% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biopax.R
\name{load_biopax}
\alias{load_biopax}
\title{Load BioPAX from Environment or OWL or Workspace File}
\usage{
load_biopax(source_name = NULL, source_dir = NULL)
}
\arguments{
\item{source_name}{Name of a BioPAX source.}

\item{source_dir}{Directory with BioPAX files (optional).}
}
\description{
Tries to load BioPAX object in several ways: searches for a BioPAX object in the environment, 
and prompts if more than one object or no object has been found. Otherwise, looks for an \code{*.owl} file in the specified directory. 
Returns a BioPAX object.
}
\author{
Ivan Grishagin
}
