% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biopax_file.R
\name{load_biopax_file}
\alias{load_biopax_file}
\title{Load BioPAX from OWL File}
\usage{
load_biopax_file(source_name = NULL, source_dir = NULL)
}
\arguments{
\item{source_name}{Name of a BioPAX source.}

\item{source_dir}{Directory with BioPAX files (optional).}
}
\description{
Tries to load BioPAX object from an OWL file.
}
\author{
Ivan Grishagin
}
