% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biopax_pathways.R
\name{load_biopax_pathways}
\alias{load_biopax_pathways}
\title{Load BioPAX Pathways From BioPAX Object}
\usage{
load_biopax_pathways(owl_biopax = NULL)
}
\arguments{
\item{owl_biopax}{A BioPAX object.}
}
\description{
Extracts all pathways from a BioPAX Object.
}
\author{
Ivan Grishagin
}
