% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_inxight_genes_per_source.R
\name{load_inxight_genes_per_source}
\alias{load_inxight_genes_per_source}
\title{Load Dataframe with Genes/Pathways}
\usage{
load_inxight_genes_per_source(source_name, toxdb_genes_file = NULL,
  all_pathways)
}
\arguments{
\item{source_name}{Name of the BioPAX source.}

\item{toxdb_genes_file}{File that has a table of all genes listed for all pathways.}

\item{all_pathways}{Dataframe with all desired pathways.}
}
\description{
Load a table with all genes listed for each of the pathways.
}
\author{
Ivan Grishagin
}
