% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_biopax_toxdb_comparison_summary.R
\name{MAIN_biopax_toxdb_comparison_summary}
\alias{MAIN_biopax_toxdb_comparison_summary}
\title{Prepare BioPAX-ToxDB Comparison Summary}
\usage{

  MAIN_biopax_toxdb_comparison_summary(work_dir = "D:/Dropbox/Rancho/NCATS/ToxDB/2016-09-08 RESULTS toxdb-vs-biopax genes comparison",
  output_dir = "./summary",
  pw_matchup_file = "../_source_files/pathways_matched_to_sources_v015.xlsx",
  pw_src_file = "../_source_files/human_pathways_rbs_curated_20150615_with_source_info.xlsx",
  gene_df_file = "./2016-09-08_ALL_SOURCES_toxdb_vs_biopax_all_genes.xlsx")
}
\arguments{
\item{work_dir}{Work directory.}

\item{output_dir}{Output directory (will be created if non-existent).}

\item{pw_matchup_file}{File with pathway matchups between BioPAX sources and ToxDB list.}

\item{pw_src_file}{ToxDB pathway spreadsheet with curation information.}

\item{gene_df_file}{Spreadsheet with BioPAX-ToxDB gene comparison.}
}
\description{
Prepares comprehensive summary of comparison of genes in BioPAX sources to ToxDB gene list.
}
\author{
Ivan Grishagin
}

