% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_compare_summarize_inxight2biopax.R
\name{MAIN_compare_summarize_inxight2biopax}
\alias{MAIN_compare_summarize_inxight2biopax}
\title{Compare and Summarize Source BioPAX files to Inxight Pathways List}
\usage{
MAIN_compare_summarize_inxight2biopax(wkdir)
}
\arguments{
\item{wkdir}{Work directory.}
}
\description{
Compare genes/proteins extracted from BioPAX to inxight pathways list, and then 
prepare a comprehensive summary such comparison.
}
\author{
Ivan Grishagin
}
