% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_compare_toxdb_biopax.R
\name{MAIN_compare_toxdb_biopax}
\alias{MAIN_compare_toxdb_biopax}
\title{Compare Proteins Extracted from BioPAX to ToxDB Proteins}
\usage{
MAIN_compare_toxdb_biopax(work_dir = NULL, source_owl_dir = NULL,
  output_dir = NULL, pw_matchup_file = "default",
  toxdb_genes_file = "default", source_name = NULL, owl_biopax = NULL,
  verbose = FALSE)
}
\arguments{
\item{work_dir}{Directory, where all source files/folders are located and will be written to.}

\item{source_owl_dir}{Directory with \code{*.owl} files. 
If no directory provided, a prompt will pop up. See \code{load.biopax} function for more information.}

\item{output_dir}{.}

\item{pw_matchup_file}{File that provides match-up between toxdb pathway names and BioPAX pathway names.}

\item{toxdb_genes_file}{File that has a table of all genes listed for all pathways.}

\item{source_name}{Name of the BioPAX source.}

\item{owl_biopax}{BioPAX object.}

\item{verbose}{Logical. Show all relevant progress messages?}
}
\description{
Main function of the package. Extract all proteins ("genes") from a BioPAX source, 
convert their IDs to gene IDs and Symbols, and then compare those to a supplied gene list.
}
\author{
Ivan Grishagin
}

