% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_biopax_files.R
\name{merge_biopax_files}
\alias{merge_biopax_files}
\title{Merge Multiple BioPAX Files into One File}
\usage{
merge_biopax_files(source_dir = NULL, change_ids = TRUE, filename = NULL)
}
\arguments{
\item{source_dir}{Directory with BioPAX files.}

\item{change_ids}{Logical. Change ids to unify them throughout the files?}

\item{filename}{Output filename (optional). If NULL (default), does not write the output to file.}
}
\description{
Merge multiple BioPAX files into one BioPAX file, while preserving all unique relationships between pathway components.
}
\author{
Ivan Grishagin
}
