% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_biopax_components.R
\name{remove_duplicate_biopax_components}
\alias{remove_duplicate_biopax_components}
\title{Remove Duplicate Components from BioPAX Object}
\usage{
remove_duplicate_biopax_components(biopax)
}
\description{
BioPAX objects frequently have components that are identical in everything, but id. 
This function removes all such duplicates, and changes all property_attr_values accordingly.
}
\author{
Ivan Grishagin
}
