\name{RIGbiopax-package}
\alias{RIGbiopax-package}
\alias{RIGbiopax}
\docType{package}
\title{
\packageTitle{RIGbiopax}
}
\description{
\packageDescription{RIGbiopax}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RIGbiopax}
\packageIndices{RIGbiopax}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RIGbiopax}

Maintainer: \packageMaintainer{RIGbiopax}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
