% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pathway_into_chunks.R
\name{split_pathway_into_chunks}
\alias{split_pathway_into_chunks}
\title{Split BioPAX Pathway into N Chunks}
\usage{
split_pathway_into_chunks(biopax, pw_id, Nchunks = 2)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pw_id}{ID of a pathway to split.}

\item{Nchuncks}{How many new BioPAX objects to produce from the original one?}
}
\description{
Split a given pathway supplied in a BioPAX object into N new BioPAX objects.
}
\author{
Ivan Grishagin
}
