% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striphash.R
\name{striphash}
\alias{striphash}
\title{Remove # from String}
\usage{
striphash(vect)
}
\arguments{
\item{vect}{String vector.}
}
\description{
Removes the first \code{#} from each element of a string vector.
}
\author{
Ivan Grishagin
}
