% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_biopax_ids.R
\name{unify_biopax_ids}
\alias{unify_biopax_ids}
\title{Uniformly Format IDs in a BioPAX-Style Data Table}
\usage{
unify_biopax_ids(biopax_dt, idtag = NULL,
  exclude_id_pattern = "inxight_pathways", exclude_class = "Pathway")
}
\arguments{
\item{biopax_dt}{BioPAX-style data table.}

\item{idtag}{String to append to each id (defaults to \code{NULL}).}

\item{exclude_id_pattern}{Exclude components with such pattern in their IDs and of class \code{exclude_class} from ID unification.}

\item{exclude_class}{Exclude components with such class and \code{exclude_id_pattern} from ID unification.}
}
\description{
Converts IDs in a BioPAX-style data table to a uniform format.
}
\details{
ID unification is conducted by appending a class of a component to its throughout number.
E.g. for a component of class Protein, #11 from the top of the table, ID will be "Protein11".
}
\author{
Ivan Grishagin
}
