% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RIGbiopax.R
\name{update_RIGbiopax}
\alias{update_RIGbiopax}
\title{Update RIGbiopax from GitHub}
\usage{
update_RIGbiopax()
}
\description{
Updates the package. It may be necessary to restart the R session after the update.
}
\author{
Ivan Grishagin
}

