% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pathways_from_biopax.R
\name{extract_pathways_from_biopax}
\alias{extract_pathways_from_biopax}
\title{Extract Pathways from BioPAX}
\usage{
extract_pathways_from_biopax(biopax, pw_ids, filename = NULL)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pw_ids}{Pathway ids to extract.}

\item{filename}{Filename of the new biopax.}
}
\description{
Extract pathways from a BioPAX object, then convert them into a separate biopax object.
}
\author{
Ivan Grishagin
}
