% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biopax_file.R
\name{load_biopax_file}
\alias{load_biopax_file}
\title{Load BioPAX from OWL File}
\usage{
load_biopax_file(source_name = NULL, source_dir = NULL)
}
\arguments{
\item{source_name}{Name of a BioPAX source.
Used to search for a directory with the source OWL file.}

\item{source_dir}{Directory with BioPAX files (optional).}
}
\description{
Tries to load a BioPAX object from an OWL file.
}
\details{
This function tries to perform a fuzzy search for a directory 
that contains \code{source_name} inside a \code{source_dir}. 
This is useful when you have several "sources" of pathways, 
i.e. SciSig, Biocarta, etc., with only one OWL file per directory.
}
\author{
Ivan Grishagin
}
