% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_biopax_toxdb_comparison_summary.R
\name{MAIN_biopax_toxdb_comparison_summary}
\alias{MAIN_biopax_toxdb_comparison_summary}
\title{Prepare BioPAX-ToxDB Comparison Summary}
\usage{
MAIN_biopax_toxdb_comparison_summary(work_dir = "default",
  output_dir = "./summary", pw_matchup_file = "default",
  pw_src_file = "default", genes_df)
}
\arguments{
\item{work_dir}{Work directory.}

\item{output_dir}{Output directory (will be created if non-existent).}

\item{pw_matchup_file}{File with pathway matchups between BioPAX sources and Inxight Pathways list.}

\item{pw_src_file}{Inxight Pathways spreadsheet with pathway curation information.}

\item{genes_df}{Either a dataframe with BioPAX-ToxDB gene comparison 
or a filename of a spreadsheet with such dataframe.}
}
\description{
Prepares comprehensive summary of comparison of genes in BioPAX sources to ToxDB gene list.
}
\author{
Ivan Grishagin
}
