% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitbiopax_from_dt.R
\name{splitbiopax_from_dt}
\alias{splitbiopax_from_dt}
\title{Convert BioPAX Object to List of BioPAX Objects (by Pathway)}
\usage{
splitbiopax_from_dt(biopax, pwtoextract = NULL, write_to_files = FALSE,
  unifyids = TRUE)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pwtoextract}{Vector of pathway ids. If \code{NULL} (default), all pathways are extracted.}

\item{write_to_files}{Logical. Write to files? Default is \code{FALSE}.}

\item{unifyids}{Logical. Unify IDs by making all elements of one class receive ids in the form "ClassNUMBER", 
with NUMBER ranging from one to number of elements of such class? Default is \code{TRUE}.}
}
\description{
Splits a BioPAX object into a list of BioPAX objects by pathway. 
Has an option of recording all of these BioPAX objects to files.
}
\details{
Returns a list of biopax objects.
}
\author{
Ivan Grishagin
}
